add `utils` tests
